Create `Ipv4NetworkSubSet` helper iterator
This is a helper type that is created with an initial IPv4 network,
represented as a `u32` representation of its IP address and a `u8` with
its prefix, and the prefix of a larger IPv4 network which contains the
initial IPv4 network. Since one is contained in the other, the IP
address shares most bits, so it only needs to keep track of it once.
However, proper care must be taken so that the initial network prefix is
larger than the larger network prefix.

The iterator will produce `Ipv4Network` items representing increasing
network ranges that are still inside the larger network but do not
include the initial network.