Create `Ipv4NetworkSubResult` helper type
This is also an iterator, and it wraps around the `Ipv4NetworkSubSet`
type. However, it also includes entries for when the minuend and the
subtrahend don't overlap and for when the subtrahend completly ovelaps
the minuend.

In the former case, there is no overlap of the networks. Therefore,
nothing is subtracted from the original network (the minuend), so the
result is just the original network.

In the latter case, since the subtrahend completly overlaps the minuend,
it is in effect completely removing the original network. Therefore, the
result is no networks at all, and the resulting iterator is empty.