Implement subtraction of multiple `Ipv4Network`s
This allows one to subtract multiple `Ipv4Network`s from a single
`Ipv4Network`. This works by having a subtraction chain, where the
original minuend `Ipv4Network` is subtracted by the first `Ipv4Network`
subtrahend, and the result is then subtracted by the second
`Ipv4Network` subtrahend, and so on.

The implementation uses dynamic dispatch because it is impossible to
know how many subtrahends there will be. This is a simple initial
implementation, and tere might be other solutions, such as creating a
custom iterator and looping around all subtrahends to provide the
results, but it quickly becomes more complicated. This might be doable
as future work.

Another possible future work is to remove the dynamic dispatch for known
sizes of subtrahend arrays, but this would probably require const
generics to be stabilized.