Create `IpNetworkSubResult` wrapper type
This type represents the result of the subtraction of one `IpNetwork`
from another. The implementation just wraps the resulting subtraction of
the protocol specific (IPv4 or IPv6) subtraction result.

Note that this means that it is not currently possible to subtract
`IpNetwork`s of different protocols (i.e., either an IPv4 network from
an IPv6 network or vice-versa).